More dropdowns

+++ app/src/main/java/com/browntowndev/liftlab/core/persistence/dtos/DropSetDto.kt
@@ -6,8 +6,8 @@ data class DropSetDto (
     override val id: Long = 0,
     override val workoutLiftId: Long,
     override val position: Int,
-    override val rpeTarget: Double,
+    override val rpeTarget: Float,
     override val repRangeBottom: Int,
     override val repRangeTop: Int,
-    val dropPercentage: Double,
+    val dropPercentage: Float,
 ) : GenericCustomLiftSet
